Increase max latency of removable drive plugin to 28 days.

+++ bramble-android/src/main/java/org/briarproject/bramble/plugin/bluetooth/AndroidBluetoothPlugin.java
@@ -87,7 +87,7 @@ class AndroidBluetoothPlugin extends
 			Clock clock,
 			Backoff backoff,
 			PluginCallback callback,
-			int maxLatency,
+			long maxLatency,
 			int maxIdleTime) {
 		super(connectionLimiter, connectionFactory, ioExecutor,
 				wakefulIoExecutor, secureRandom, backoff, callback,
